containers/ml_pipeline_worker/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pipeline-worker"
version = "0.1.0"
description = "Self-contained ML pipeline tasks for running inside DockerOperator containers."
authors = [{name = "MLOps Team"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.33",
    "mlflow>=2.7",
    "pandas>=2.1",
    "pyarrow>=12.0",
    "python-dotenv>=1.0",
    "scikit-learn>=1.3",
]

[tool.setuptools.packages.find]
where = ["src"]
